Refactored the gallery to have two carousels that interact with each other
